+++ CHANGELOG.md
@@ -19,6 +19,8 @@ into pre-release sections below.
 
 # 2.7.0a6 (unreleased)
 
+* Database:
+    * Verify task-group non-duplication constraint in `2.7.0` data-migration script (\#1927).
 * Runner:
     * Do not create local folders with 755 permissions unless `FRACTAL_BACKEND_RUNNER="slurm"` (\#1923).
 * Testing: